requirements.txt
streamlit
torch
torchvision
pillow
joblib
transformers